Relocated mod methods to API
