+++ src/main/java/io/redspace/ironsspellbooks/entity/spells/dragon_breath/DragonBreathPool.java
@@ -1,7 +1,7 @@
 package io.redspace.ironsspellbooks.entity.spells.dragon_breath;
 
 import io.redspace.ironsspellbooks.damage.DamageSources;
-import io.redspace.ironsspellbooks.entity.spells.AoeEntity;
+import io.redspace.ironsspellbooks.api.entity.spells.AoeEntity;
 import io.redspace.ironsspellbooks.damage.ISSDamageTypes;
 import io.redspace.ironsspellbooks.registries.EntityRegistry;
 import net.minecraft.core.particles.ParticleOptions;
