+++ src/main/java/io/redspace/ironsspellbooks/entity/spells/flame_strike/FlameStrike.java
@@ -1,6 +1,6 @@
 package io.redspace.ironsspellbooks.entity.spells.flame_strike;
 
-import io.redspace.ironsspellbooks.entity.spells.AoeEntity;
+import io.redspace.ironsspellbooks.api.entity.spells.AoeEntity;
 import io.redspace.ironsspellbooks.registries.EntityRegistry;
 import net.minecraft.core.particles.ParticleOptions;
 import net.minecraft.network.protocol.Packet;
