+++ src/main/java/io/redspace/ironsspellbooks/spells/TargetAreaCastData.java
@@ -1,7 +1,6 @@
 package io.redspace.ironsspellbooks.spells;
 
-import io.redspace.ironsspellbooks.entity.spells.target_area.TargetedAreaEntity;
-import net.minecraft.world.entity.Entity;
+import io.redspace.ironsspellbooks.api.entity.spells.target_area.TargetedAreaEntity;
 import net.minecraft.world.phys.Vec3;
 
 public class TargetAreaCastData extends EntityCastData {
