+++ src/main/java/io/redspace/ironsspellbooks/spells/TargetedTargetAreaCastData.java
@@ -1,10 +1,8 @@
 package io.redspace.ironsspellbooks.spells;
 
-import io.redspace.ironsspellbooks.api.spells.ICastData;
 import io.redspace.ironsspellbooks.capabilities.magic.CastTargetingData;
-import io.redspace.ironsspellbooks.entity.spells.target_area.TargetedAreaEntity;
+import io.redspace.ironsspellbooks.api.entity.spells.target_area.TargetedAreaEntity;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.phys.Vec3;
 
 public class TargetedTargetAreaCastData extends CastTargetingData {
     final TargetAreaCastData targetAreaCastData;
